# griffin-hub/openapi-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "griffin-runner API",
    "description": "API for managing and running test plans",
    "version": "1.0.0"
  },
  "components": {
    "securitySchemes": {},
    "schemas": {
      "TestPlanV1": {
        "type": "object",
        "required": ["id", "name", "version", "environment", "nodes", "edges"],
        "properties": {
          "organization": {
            "type": "string"
          },
          "project": {
            "type": "string"
          },
          "locations": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "version": {
            "type": "string",
            "enum": ["1.0"]
          },
          "frequency": {
            "$ref": "#/components/schemas/Frequency"
          },
          "environment": {
            "type": "string",
            "default": "default"
          },
          "nodes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Node"
            }
          },
          "edges": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Edge"
            }
          }
        },
        "title": "TestPlanV1"
      },
...
      "Node": {
        "title": "Node",
        "oneOf": [
          {
            "$ref": "#/components/schemas/Endpoint"
          },
          {
            "$ref": "#/components/schemas/Wait"
          },
          {
            "$ref": "#/components/schemas/Assertions"
          }
        ]
      },
      "Edge": {
        "type": "object",
        "required": ["from", "to"],
        "properties": {
          "from": {
            "type": "string"
          },
          "to": {
            "type": "string"
          }
        },
        "title": "Edge"
      },
      "Frequency": {
        "type": "object",
        "required": ["every", "unit"],
        "properties": {
          "every": {
            "type": "number"
          },
          "unit": {
            "anyOf": [
              {
                "type": "string",
                "enum": ["MINUTE"]
              },
              {
                "type": "string",
                "enum": ["HOUR"]
              },
              {
                "type": "string",
                "enum": ["DAY"]
              }
            ]
          }
        },
        "title": "Frequency"
      },
      "Endpoint": {
        "type": "object",
        "required": ["id", "type", "method", "path", "base", "response_format"],
        "properties": {
          "id": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": ["ENDPOINT"]
          },
          "method": {
            "$ref": "#/components/schemas/HttpMethod"
          },
          "path": {
            "type": "string"
          },
          "base": {
            "type": "object",
            "required": ["type", "key"],
            "properties": {
              "type": {
                "type": "string",
                "enum": ["target"]
              },
              "key": {
                "type": "string"
              }
            }
          },
          "headers": {
            "type": "object",
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "object",
                  "required": ["$secret"],
                  "properties": {
                    "$secret": {
                      "type": "object",
                      "required": ["provider", "ref"],
                      "properties": {
                        "provider": {
                          "type": "string"
                        },
                        "ref": {
                          "type": "string"
                        },
                        "version": {
                          "type": "string"
                        },
                        "field": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              ]
            }
          },
          "body": {},
          "response_format": {
            "$ref": "#/components/schemas/ResponseFormat"
          }
        },
        "title": "Endpoint"
      },
      "HttpMethod": {
        "anyOf": [
          {
            "type": "string",
            "enum": ["GET"]
          },
          {
            "type": "string",
            "enum": ["POST"]
          },
          {
            "type": "string",
            "enum": ["PUT"]
          },
          {
            "type": "string",
            "enum": ["DELETE"]
          },
          {
            "type": "string",
            "enum": ["PATCH"]
          },
          {
            "type": "string",
            "enum": ["HEAD"]
          },
          {
            "type": "string",
            "enum": ["OPTIONS"]
          },
          {
            "type": "string",
            "enum": ["CONNECT"]
          },
          {
            "type": "string",
            "enum": ["TRACE"]
          }
        ],
        "title": "HttpMethod"
      },
      "ResponseFormat": {
        "anyOf": [
          {
            "type": "string",
            "enum": ["JSON"]
          },
          {
            "type": "string",
            "enum": ["XML"]
          },
          {
            "type": "string",
            "enum": ["TEXT"]
          }
        ],
        "title": "ResponseFormat"
      },
      "Wait": {
        "type": "object",
        "required": ["id", "type", "duration_ms"],
        "properties": {
          "id": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": ["WAIT"]
          },
          "duration_ms": {
            "type": "number"
          }
        },
        "title": "Wait"
      },
      "Assertions": {
        "type": "object",
        "required": ["id", "type", "assertions"],
        "properties": {
          "id": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": ["ASSERTION"]
          },
          "assertions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Assertion"
            }
          }
        },
        "title": "Assertions"
      },
      "Assertion": {
        "anyOf": [
          {
            "allOf": [
              {
                "type": "object",
                "required": ["nodeType"],
                "properties": {
                  "nodeType": {
                    "type": "string",
                    "enum": ["JSON"]
                  }
                }
              },
              {
                "type": "object",
                "required": ["nodeId", "accessor", "path", "predicate"],
                "properties": {
                  "nodeId": {
                    "type": "string"
                  },
                  "accessor": {
                    "anyOf": [
                      {
                        "type": "string",
                        "enum": ["body"]
                      },
                      {
                        "type": "string",
                        "enum": ["headers"]
                      },
                      {
                        "type": "string",
                        "enum": ["status"]
                      }
                    ]
                  },
                  "path": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "predicate": {
                    "anyOf": [
                      {
                        "anyOf": [
                          {
                            "type": "string",
                            "enum": ["IS_NULL"]
                          },
                          {
                            "type": "string",
                            "enum": ["IS_NOT_NULL"]
                          },
                          {
                            "type": "string",
                            "enum": ["IS_TRUE"]
                          },
                          {
                            "type": "string",
                            "enum": ["IS_FALSE"]
                          },
                          {
                            "type": "string",
                            "enum": ["IS_EMPTY"]
                          },
                          {
                            "type": "string",
                            "enum": ["IS_NOT_EMPTY"]
                          }
                        ]
                      },
                      {
                        "type": "object",
                        "required": ["expected", "operator"],
                        "properties": {
                          "expected": {},
                          "operator": {
                            "$ref": "#/components/schemas/BinaryPredicateOperator"
                          }
                        }
                      }
                    ]
                  }
                }
              }
            ]
          },
          {
            "allOf": [
              {
                "type": "object",
                "required": ["nodeType"],
                "properties": {
                  "nodeType": {
                    "type": "string",
                    "enum": ["XML"]
                  }
                }
              },
              {
                "type": "object",
                "required": ["path", "expected"],
                "properties": {
                  "path": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "expected": {}
                }
              }
            ]
          },
          {
            "allOf": [
              {
                "type": "object",
                "required": ["nodeType"],
                "properties": {
                  "nodeType": {
                    "type": "string",
                    "enum": ["TEXT"]
                  }
                }
              },
              {
                "type": "object",
                "required": ["path", "expected"],
                "properties": {
                  "path": {
                    "type": "string"
                  },
                  "expected": {}
                }
              }
            ]
          }
        ],
        "title": "Assertion"
      },
      "BinaryPredicateOperator": {
        "anyOf": [
          {
            "type": "string",
            "enum": ["EQUAL"]
          },
          {
            "type": "string",
            "enum": ["NOT_EQUAL"]
          },
          {
            "type": "string",
            "enum": ["GREATER_THAN"]
          },
          {
            "type": "string",
            "enum": ["LESS_THAN"]
          },
          {
            "type": "string",
            "enum": ["GREATER_THAN_OR_EQUAL"]
          },
          {
            "type": "string",
            "enum": ["LESS_THAN_OR_EQUAL"]
          },
          {
            "type": "string",
            "enum": ["CONTAINS"]
          },
          {
            "type": "string",
            "enum": ["NOT_CONTAINS"]
          },
          {
            "type": "string",
            "enum": ["STARTS_WITH"]
          },
          {
            "type": "string",
            "enum": ["ENDS_WITH"]
          },
          {
            "type": "string",
            "enum": ["NOT_STARTS_WITH"]
          },
          {
            "type": "string",
            "enum": ["NOT_ENDS_WITH"]
          }
        ],
        "title": "BinaryPredicateOperator"
      }
    }
  },
  "paths": {
...
    "/plan/": {
      "post": {
        "tags": ["plan"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TestPlanV1"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["data"],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/TestPlanV1"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["error"],
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["error"],
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["error"],
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["error"],
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["error"],
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "502": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["error"],
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "503": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["error"],
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "504": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["error"],
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["plan"],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "projectId",
            "required": false
          },
          {
            "schema": {
              "type": "number",
              "minimum": 1,
              "maximum": 100
            },
            "in": "query",
            "name": "limit",
            "required": false
          },
          {
            "schema": {
              "type": "number",
              "minimum": 0
            },
            "in": "query",
            "name": "offset",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["data", "total", "page", "limit"],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/TestPlanV1"
                      }
                    },
                    "total": {
                      "type": "number"
                    },
                    "page": {
                      "type": "number"
                    },
                    "limit": {
                      "type": "number"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["error"],
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["error"],
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["error"],
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["error"],
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["error"],
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "502": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["error"],
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "503": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["error"],
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "504": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["error"],
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
...
}